Address validation and checksum (#151)
Added validation to Address constructor: an invalid string throws an IllegalArgumentException.

Address.toString() returns the Ethereum address in checksum case (good format for the user interfaces etc.) In protocol level we sometimes need the address in lowercase format. That's why the internal format is in lowercase, and there is a toLowerCaseString() accessor to get that format efficiently.